requirements.txt
pyspark==3.5.0
boto3==1.34.0
pandas==2.2.0
matplotlib==3.8.0
seaborn==0.13.0
python-dotenv==1.0.0
jupyter==1.0.0